Create VISUALIZATION
INSIGHT 1.
 https://public.tableau.com/app/profile/faith.idahota/viz/AveragePovertyRatebyStateUS201 5/c

SUMMARY
 The bar chart shows the average poverty rate of each state in the United States in 2015. Puerto Rico had the highest (49.37%), followed by Mississippi with (25.72%). Connecticut had the lowest(9.40%).

DESIGN
 I used a bar chart because positions are easy to see and blue was used in order not to distract the user with colors. 

INSIGHT 2
https://public.tableau.com/app/profile/faith.idahota/viz/TheRelationshipBetweenAveragePovertyandAverageunemploymentbyState/TheRelationshipBetweenAveragePovertyandAverageunemploymentbyState

SUMMARY 
The scatterplot shows the relationship between the average poverty rate and average unemployment rate in the United States in 2015. The higher the unemployment rate, the higher the poverty rate. Puerto Rico and Mississippi had the highest unemployment and poverty rates and Nebraska had the lowest poverty and unemployment rate.

DESIGN 
Scatterplot was used because it helped to visualize the relationship between two numerical variables; colors were used in order to easily distinguish one state from another.

DASHBOARD
https://public.tableau.com/app/profile/faith.idahota/viz/OverviewofthePopulationDensityandthePerCapitaIncomeofStatesinUS2015/OverviewofthePopulationDensityandthePerCapitaIncomeofStatesinUS2015

The dashboard shows two insights ( the population density and the average income per capita of each state) making it easy for users to quickly glance through without having to navigate through individual charts.

DESIGN
Dashboard was used because of its ability to combine insights to give a broader view.

SUMMARY 
The map shows the population and the average income per capita of each state in 2015 while the area plot shows the average income per capita of the states. California has the highest population (38,421,464) with per capita income of (27,903), followed by Texas (26,538,614) with per capita income of (23,556). Wyoming(577,679) had the lowest population and a per capita income of (29,224). The District of Columbia had the highest per capita income of (47,675) with population density of (647,484). It seemed an increase in population resulted in a decrease in per capita income.

DESIGN
Maps help to quickly get an overview of the states. Colors were used to differentiate the states. Area plot was used in descending order for users to quickly get the state with the highest to the lowest.